Add Source::KeyAndPattern and DIDMethods::generate
Allow passing an additional string to a DID method in the generate
function.

Implement generate on DIDMethods to allow generating a DID from a set of
DID methods given a DID method name and optional additional string,
separated by a colon (':').

The additional string may be a prefix or pattern for the DID's
method-specific ID, according to the DID method.